fixing multilimiter moving packages
